+++ flowserv/app/result.py
@@ -1,0 +1,24 @@
+# This file is part of the Reproducible and Reusable Data Analysis Workflow
+# Server (flowServ).
+#
+# Copyright (C) 2019-2020 NYU.
+#
+# flowServ is free software; you can redistribute it and/or modify it under the
+# terms of the MIT License; see LICENSE file for more details.
+
+"""Classes that wrap serialized run result dictionaries. The aim is to make
+it easier for a developer that uses the flowserv application object to access
+the results and resources of workflow runs.
+"""
+
+
+class ResultFile(object):
+    """
+    """
+    pass
+
+
+class RunResult(object):
+    """
+    """
+    pass