+++ changelog.md
@@ -63,6 +63,10 @@
 * Remove old workflow runs from database (\#9).
 
 
-### 0.4.1 - 2020-09-17
+### 0.5.0 - 2020-09-24
 
-* Allow user-defined workflow and group identifier (\#59)
+* Allow user-defined workflow and group identifier (\#59).
+* Merge `flowserv` and `flowapp` console scripts.
+* User-defined keys for output files (\#60).
+* Add functionality for asynchronous runs to flowserv application object.
+* Change flowserv application run result object (\#61).